fix(math): enforce np.float64 dtype during array creation
Explicitly defined `dtype=np.float64` in `RiskEngine` and `QuantEngine` where `np.array()` was previously relying on implicit typecasting.

This prevents a known bug where an initial input list of only integers causes `np.array` to infer an integer dtype. Subsequent floating-point math on the array (e.g., float division for weights) would then crash with a `_UFuncOutputCastingError`.

Also appended a strategic suggestion to `future/suggestions.MD` to standardize this practice across all financial operations.

+++ backend/quant_engine.py
@@ -427,7 +427,7 @@ class QuantEngine:
                 
                 # Simple alignment: assume same length and timestamps for now
                 # In production, we'd use a more robust outer-join and interpolation
-                closes = np.array([float(item.close) for item in series])
+                closes = np.array([float(item.close) for item in series], dtype=np.float64)
                 
                 if len(closes) < seq_len:
                     # Pad if necessary or truncate
@@ -464,7 +464,7 @@ class QuantEngine:
             
             # 4. Calculate Portfolio CVaR for the optimized portfolio
             # Portfolio returns = ReturnsMatrix * weights
-            w_arr = np.array([float(w) for w in weights])
+            w_arr = np.array([float(w) for w in weights], dtype=np.float64)
             portfolio_returns = np.dot(returns_matrix, w_arr)
             
             from utils.risk_engine import RiskEngine
